Add modules for alsa volume and backlight brighness
These modules are special, in that they are only shown for a short
period of time, after the backlight_trigger and volume_trigger functions
have been called. That way the bar is not cluttered with information,
which doesn't have to be visible all the time.

The `mod_separator` variable was removed in favour of adding separators
inside `mod_format` on a per mod basis. That way the two new modules
won't consume any space in their hidden state. Otherwise a module
without output would still have a separator displayed.